Debugger: Don't kill lldb process if we get only a warning

+++ src/plugins/debugger/lldb/lldbengine.cpp
@@ -820,7 +820,8 @@ void LldbEngine::readLldbStandardError()
     qDebug() << "\nLLDB STDERR" << err;
     //qWarning() << "Unexpected lldb stderr:" << err;
     showMessage(_(err), LogError);
-    m_lldbProc.kill();
+    if (!err.startsWith("warning:"))
+        m_lldbProc.kill();
 }
 
 void LldbEngine::readLldbStandardOutput()
